user 보유 캐시 int -> float 변경

+++ src/main/java/bit/bitgroundspring/dto/response/UserAssetResponse.java
@@ -13,6 +13,6 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserAssetResponse {
-    private Integer cash;
+    private Float cash;
     private List<UserAssetProjection> userAssets;
 }
